requirements.txt
# Web Framework
flask>=3.0.0
gunicorn>=23.0.0

# OpenAI
openai>=1.0.0

# LangGraph / LangChain (Module 3)
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0

# Slack (for human-in-the-loop)
slack-bolt>=1.18.0
slack-sdk>=3.21.0

# Google APIs (Sheets, Gmail)
google-api-python-client>=2.100.0
google-auth>=2.20.0

# Salesforce
simple-salesforce>=1.12.0

# HTTP Requests
requests>=2.31.0

# Environment Variables
python-dotenv>=1.0.0
